move assertions to separate explorer_tests.py file
